third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/w3c/webperf/approved/navigation-timing/html/test_timing_server_redirect-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.navigation is defined 
PASS timing.navigation.type is TYPE_NAVIGATE 
PASS navigation.redirectCount == 1 on an server redirected navigation 
PASS window.performance.timing.navigationStart > 0 
PASS window.performance.timing.redirectStart > 0 
PASS window.performance.timing.redirectStart >= window.performance.timing.navigationStart 
PASS window.performance.timing.redirectEnd > 0 
PASS window.performance.timing.redirectEnd >= window.performance.timing.redirectStart 
PASS window.performance.timing.fetchStart > 0 
PASS window.performance.timing.fetchStart >= window.performance.timing.redirectEnd 
Harness: the test ran to completion.
